apfs: Make mode in apfs_inode u16
A file's mode is 16 Bit only, so no need to waste another 16 Bit for
it in struct apfs_inode.

+++ apfs.h
@@ -279,7 +279,7 @@ struct apfs_info {
 struct apfs_inode {
 	struct inode			vfs_inode;
 
-	u32				mode;
+	u16				mode;
 	u32				nlink;
 	u32				uid;
 	u32				gid;
